🔧 refactor: 타입 정의를 MyExperiences.ts로 이동
